requirements.txt
pandas 
sqlalchemy
python-dotenv